REF: Move interfaces from misc to cmtklib/diffusion
