+++ tests/test_dipy_pft.py
@@ -4,7 +4,7 @@ from os import path as op
 
 
 def test_ExtractPVEsFrom5TT(subject, session, in_5tt_image, ref_image, base_dir):
-    from cmtklib.interfaces.misc import ExtractPVEsFrom5TT
+    from cmtklib.diffusion import ExtractPVEsFrom5TT
     from nipype import Node
 
     pve_extracter = Node(interface=ExtractPVEsFrom5TT(),
